+++ src/sll/include/sll/location.h
@@ -65,10 +65,10 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_file_offset_t sll_get_location(const sll_a
  * \name sll_get_name
  * \group location
  * \desc Docs!
- * \arg const sll_object_t* v
+ * \arg sll_object_t* v
  * \arg sll_string_t* o
  */
-__SLL_EXTERNAL void sll_get_name(const sll_object_t* v,sll_string_t* o);
+__SLL_EXTERNAL void sll_get_name(sll_object_t* v,sll_string_t* o);