+++ src/sll/include/sll/operator.h
@@ -137,10 +137,10 @@ __SLL_EXTERNAL void sll_operator_assign_range_step(sll_object_t* a,sll_object_t*
  * \name sll_operator_bool
  * \group operator
  * \desc Docs!
- * \arg const sll_object_t* a
+ * \arg sll_object_t* a
  * \ret sll_bool_t
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_operator_bool(const sll_object_t* a);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_operator_bool(sll_object_t* a);
 
 
 
@@ -162,11 +162,11 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_operator_cast(sll_object_t* 
  * \name sll_operator_compare
  * \group operator
  * \desc Docs!
- * \arg const sll_object_t* a
- * \arg const sll_object_t* b
+ * \arg sll_object_t* a
+ * \arg sll_object_t* b
  * \ret sll_compare_result_t
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_compare_result_t sll_operator_compare(const sll_object_t* a,const sll_object_t* b);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_compare_result_t sll_operator_compare(sll_object_t* a,sll_object_t* b);
 
 
 
@@ -213,11 +213,11 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_operator_copy(sll_object_t* 
  * \name sll_operator_equal
  * \group operator
  * \desc Docs!
- * \arg const sll_object_t* a
- * \arg const sll_object_t* b
+ * \arg sll_object_t* a
+ * \arg sll_object_t* b
  * \ret sll_bool_t
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_operator_equal(const sll_object_t* a,const sll_object_t* b);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_operator_equal(sll_object_t* a,sll_object_t* b);
 
 
 
@@ -353,11 +353,11 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_operator_shr(sll_object_t* a
  * \name sll_operator_strict_equal
  * \group operator
  * \desc Docs!
- * \arg const sll_object_t* a
- * \arg const sll_object_t* b
+ * \arg sll_object_t* a
+ * \arg sll_object_t* b
  * \ret sll_bool_t
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_operator_strict_equal(const sll_object_t* a,const sll_object_t* b);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_operator_strict_equal(sll_object_t* a,sll_object_t* b);